backend/requirements.txt
fastapi==0.104.1
pydantic==2.4.2
pymongo==4.6.0
uvicorn==0.23.2
rdflib==7.0.0
requests==2.31.0
python-dotenv==1.0.0
PyLD==2.0.3
Levenshtein==0.23.0
